Updated assignment of UI-Session9

+++ UI/UI-Session9/SumUsingES6/src/app.js
@@ -4,8 +4,8 @@ const result = document.getElementById('result');
 
 result.addEventListener("click", function(){
 
-	const firstNumber = parseInt(document.getElementById("firstNumber").value,10);
-	const secondNumber = parseInt(document.getElementById("secondNumber").value,10);
+	const firstNumber = parseFloat(document.getElementById("firstNumber").value,10);
+	const secondNumber = parseFloat(document.getElementById("secondNumber").value,10);
 	const resultAdd = document.getElementById('resultAdd');
 
 	resultAdd.textContent = `The sum of ${firstNumber} and ${secondNumber} is ${add(firstNumber)(secondNumber)}.`;
